\name{ISCCNBScolours}
\docType{data}
\alias{ISCCNBScolours}
\alias{ISCCNBScolors}
\title{
  An ISCC-NBS Colour List
}
\description{
  A colour list based on the ISCC-NBS System of Colour Designation.
  There are 267 colour names, with each name corresponding to a
  "block" or region of Munsell colour space.  The sRGB colour
  for each name in the list represents the centroid of the relevant
  block.
}
\usage{
ISCCNBScolours
}
\format{ A \code{"colourList"} object. }
\references{
}
\keyword{ datasets }% use one of  RShowDoc("KEYWORDS")
